Revert "Admin View (#7)"
This reverts commit 5b13681711ddefe0a842d13444ced5da4ffe36d7.
