+++ src/actions/index.js
@@ -75,6 +75,7 @@ export function getProducts(products) {
     };
 }
 export function updateProducts(products) {
+    console.log('updateProducts action', products);
     return {
         type: UPDATE_PRODUCTS,
         products
