+++ src/components/controller/Home/index.js
@@ -3,6 +3,7 @@ import { openModal, toggleShowDescription, updateQuantity } from 'actions';
 import Home from 'components/view/Home';
 
 const mapStateToProps = (state) => {
+
     return {
         products: state.products,
         userRole: state.user.role,
