--- src/components/view/Admin/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const AdminPage = props => {
-    return (
-        <div className="page">ADMIN VIEW</div>
-    );
-};
-
-export default AdminPage;
