+++ src/components/view/Authorized/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AuthorizedPage = props => {
+    return (
+        <div className="page">This is an Authorized Page</div>
+    );
+};
+
+export default AuthorizedPage;
